code clean; added WriteMemory function; bug fix;

+++ ReadWriteMemory/Models/MemoryAddress.cs
@@ -2,14 +2,14 @@
 
 public sealed class MemoryAddress
 {
-    public MemoryAddress(int address, string moduleName = "", params int[]? offsets)
+    public MemoryAddress(long address, string moduleName = "", params int[]? offsets)
     {
         Address = address;
         ModuleName = moduleName;
         Offsets = offsets;
     }
 
-    public int Address { get; }
+    public long Address { get; }
     public string ModuleName { get; }
     public int[]? Offsets { get; }
 }
